Use correct #include syntax
Double quote syntax should be used for local includes. Although either syntax works when the library is installed normally, the incorrect use of angle braces sytax to #include local files causes compilation to fail when the library is installed elsewhere. Most commonly this is done to bundle library dependencies with a sketch to distribute it as a self contained package.